+++ CHANGELOG.md
@@ -5,6 +5,7 @@ v0.4 (TBD)
  - Random number generator functions: `vrand` and `frand`
  - Optionally set DICOM prefix and entropy source in "Advanced"
  - Keep cross-file UID connections by updating "Referenced" tags
+ - Value functions no longer have an index parameter (parent DICOM elements don't have a value)
 
 v0.3 (2020.03.31)
 --------------------